chore: move tests and fix them
